\name{preybiom}
\alias{preybiom}
\docType{data}
\title{  Jackal and Genet diet in Algeria }
\description{
  This data set gives the results of dietary analysis performed by Amroun Mansour in two sites of Kabylie, Algeria
}
\usage{data(preybiom)}
\format{
  A data frame with 2196 observations on the following variables.
  \describe{
    \item{\code{faeces}}{a factor for faeces corresponding to faeces identification numbers}
    \item{\code{site}}{a factor for study sites with levels \code{S} Sebaou \code{Y} Yacouren }
    \item{\code{saison}}{a factor for seasons with levels \code{H} \code{HD} \code{HP} \code{S} \code{SD} \code{SP}}
    \item{\code{sp}}{a factor for species with levels \code{C} Jackal \code{G} Genet }
    \item{\code{category}}{a factor for dietary items with levels \code{dech} \code{ind} \code{ins} \code{mam} \code{mol} \code{oisauv} \code{oisdom} \code{rept} \code{vege} \code{vegn}}
    \item{\code{biomasse}}{a numeric vector for the weight of each dietary item}
  }
}

\references{
  M. Amroun, P. Giraudoux and P. Delattre 2006 Comparative study of the diets of two sympatric carnivores - the Jackal (Canis aureus) and the Genet (Genetta genetta) - at two sites in Kabylia, Algeria. Mammalia, 70 (3/4): 247-254.
}

\keyword{datasets}
